requirements.txt
newspaper3k>=0.2.8,<0.3.0
spacy>=3.8,<4.0
streamlit>=1.45,<2.0
wordcloud>=1.9,<2.0
pandas>=2.2,<3.0
matplotlib>=3.9,<4.0
lxml_html_clean>=0.4,<1.0
nltk>=3.9,<4.0

